core::rt: Move uv idle tests to idle mod
